Prevent store recreation every hot update
In order to the fix the "<Provider> does not support changing store on
the fly" error, I added the configureStore function and the hot module
update listener for reducers. This caused a problem where each time
<App> rerendered, the store got recreated.

See https://goo.gl/MSWdCr for the check that caused the <Provider>
error. Moving store creation to index.js (which sets up hot reloading)
fixes this issue.